+++ neurolife/neurolife/creature.h
@@ -7,18 +7,18 @@ class Creature
 public:
 	Creature(Field * filed);
 
-	Creature(Field * filed, size_t x, size_t y);
+	Creature(Field * filed, double x, double y);
 
 	void kill() { is_alive = false; }
 	bool is_ok() {return is_alive; }
 
-	size_t get_x() const { return x; }
-	size_t get_y() const { return y; }
+	double get_x() const { return x; }
+	double get_y() const { return y; }
 
 protected:
 
-	size_t x;
-	size_t y;
+	double x;
+	double y;
 
 	bool is_alive;
